requirements.txt
python-telegram-bot==20.3
pg8000==1.29.6
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.0.2
openpyxl==3.1.1
